@@ -8,7 +8,7 @@ using FoundationaLLM.Common.OpenAPI;
 using Microsoft.AspNetCore.Http.Features;
 using Microsoft.Extensions.Configuration.AzureAppConfiguration;
 using Microsoft.Extensions.Options;
-using Microsoft.OpenApi.Models;
+using Microsoft.OpenApi;
 using Swashbuckle.AspNetCore.SwaggerGen;
 
 namespace FoundationaLLM.Management.API
@@ -148,19 +148,9 @@ namespace FoundationaLLM.Management.API
                     // Integrate xml comments
                     options.IncludeXmlComments(filePath);
 
-                    options.AddSecurityRequirement(new OpenApiSecurityRequirement
+                    options.AddSecurityRequirement(document => new OpenApiSecurityRequirement
                     {
-                        {
-                            new OpenApiSecurityScheme
-                            {
-                                Reference = new OpenApiReference
-                                {
-                                    Id = "azure_auth",
-                                    Type = ReferenceType.SecurityScheme
-                                }
-                            },
-                            new[] {"user_impersonation"}
-                        }
+                        [new OpenApiSecuritySchemeReference("azure_auth", document)] = ["user_impersonation"]
                     });
 
                     options.AddSecurityDefinition("azure_auth", new OpenApiSecurityScheme
@@ -206,7 +196,7 @@ namespace FoundationaLLM.Management.API
                     => await Results.Problem().ExecuteAsync(context)));
 
             // Configure the HTTP request pipeline.
-            app.UseSwagger(p => p.SerializeAsV2 = true);
+            app.UseSwagger(options => options.OpenApiVersion = OpenApiSpecVersion.OpenApi2_0);
             app.UseSwaggerUI(
                 options =>
                 {
